consolidate `metrics/2` in an `Admin` module that can be used for other admin functions such as “ping”
